Separate reviewposts into new file and remove comments
